Assess Stock Status Based on Calculated Parameters

This function assesses the stock status based on parameters calculated by the FishPar function.

FishSS(data, LM_ratio, Pobj, Pmat, Popt)

Arguments

data

A data frame containing the necessary columns for stock status calculation.

LM_ratio

A numeric value representing the length at maturity ratio.

Pobj

A numeric value representing the percentage objective.

Pmat

A numeric value representing the percentage of mature fish.

Popt

A numeric value representing the percentage of optimally sized fish.

Value

A numeric vector containing TSB40 and LSB25.

Examples

data <- data.frame(Tx = 1:10, A = rnorm(10), B = rnorm(10), C = rnorm(10), D = rnorm(10),
                   E = rnorm(10), F = rnorm(10), G = rnorm(10), H = rnorm(10), I = rnorm(10), J = rnorm(10))
FishSS(data, 0.75, 100, 30, 25)
#> $Target_Cols
#>    Tx            A           C
#> 1   1 -0.347088466 -0.84524782
#> 2   2 -0.098550690 -0.96157149
#> 3   3  0.034766060  1.01749105
#> 4   4  0.386127022 -1.49605374
#> 5   5  0.020831228 -1.18481873
#> 6   6  0.007586777  0.63023437
#> 7   7  0.930844030  2.10125251
#> 8   8 -0.684749941 -0.61373681
#> 9   9  0.337401513 -1.63463827
#> 10 10 -0.412137704 -0.01044112
#> 
#> $Target_value
#> [1] 30
#> 
#> $Colesest_value
#> [1] 10
#> 
#> $StockStatus
#>       TSB40       LSB25 
#> -0.41213770 -0.01044112 
#>